Temporary solution and add clangd to new format.
We are in a limbo right now. lspconfig API is deprecated but its default
configurations are not yet in an lsp/*.lua format for vim.lsp.config in
neither lspconfig or in neovim itself. The migration is currently on the
way https://github.com/neovim/nvim-lspconfig/issues/3494. For now do
myself what the migration proposes. That is, copy-paste the configs in
lspconfig and write them, write them in a format suitable for
vim.lsp.config, and include them in lsp/*.lua. The configs in lspconfig
are "semi-frozen" according to the issue, so this should be fine (not
much fear of having to mantain changes myself since changes are not
expected). Wait until the configs are in the lua/ of either lspconfig or
nvim itself. If in lspconfig, replace the manual copy-paste-edits that
I'm doing with the ones they publish (should be more robust). If in
neovim itself, replace the big copy-paste-edits with smaller configs
that just override what I want them to override (I guess
vim.tbl_deep_extend applies here?); I hope that this is the way they go.